# Improve Termmap Indexing Performance +~30%