--- a/stacker/src/fastcpy.rs
+++ b/stacker/src/fastcpy.rs
@@ -75,6 +75,8 @@ fn short_copy(src: &[u8], dst: &mut [u8]) {
 
 #[inline]
 fn double_copy_trick<const SIZE: usize>(src: &[u8], dst: &mut [u8]) {
+    debug_assert!(src.len() >= SIZE);
+    debug_assert!(dst.len() >= SIZE);
     dst[0..SIZE].copy_from_slice(&src[0..SIZE]);
     dst[src.len() - SIZE..].copy_from_slice(&src[src.len() - SIZE..]);
 }