This contains many small changes to improve Termmap performance.
Most notably:
* Specialized byte compare and equality versions, instead of glibc calls.
* ExpUnrolledLinkedList to not contain inline items.

Allow compare hash only via a feature flag compare_hash_only:
64bits should be enough with a good hash function to compare strings by
their hashes instead of comparing the strings. Disabled by default

Benchmark without `compare_hash_only`
```bash
CreateHashMap/alice/174693
                        time:   [642.23 µs 643.80 µs 645.24 µs]
                        thrpt:  [258.20 MiB/s 258.78 MiB/s 259.41 MiB/s]
                 change:
                        time:   [-14.429% -13.303% -12.348%] (p = 0.00 < 0.05)
                        thrpt:  [+14.088% +15.344% +16.862%]
                        Performance has improved.
CreateHashMap/alice_expull/174693
                        time:   [877.03 µs 880.44 µs 884.67 µs]
                        thrpt:  [188.32 MiB/s 189.22 MiB/s 189.96 MiB/s]
                 change:
                        time:   [-26.460% -26.274% -26.091%] (p = 0.00 < 0.05)
                        thrpt:  [+35.301% +35.637% +35.981%]
                        Performance has improved.
CreateHashMap/numbers_zipf/8000000
                        time:   [9.1198 ms 9.1573 ms 9.1961 ms]
                        thrpt:  [829.64 MiB/s 833.15 MiB/s 836.57 MiB/s]
                 change:
                        time:   [-35.229% -34.828% -34.384%] (p = 0.00 < 0.05)
                        thrpt:  [+52.403% +53.440% +54.390%]
                        Performance has improved.
```

Benchmark with `compare_hash_only`
```bash
CreateHashMap/alice/174693
                        time:   [503.08 µs 505.07 µs 507.22 µs]
                        thrpt:  [328.46 MiB/s 329.86 MiB/s 331.16 MiB/s]
                 change:
                        time:   [-37.334% -36.881% -36.399%] (p = 0.00 < 0.05)
                        thrpt:  [+57.231% +58.432% +59.576%]
                        Performance has improved.
CreateHashMap/alice_expull/174693
                        time:   [863.67 µs 873.21 µs 883.15 µs]
                        thrpt:  [188.64 MiB/s 190.79 MiB/s 192.90 MiB/s]
                 change:
                        time:   [-35.224% -34.352% -33.461%] (p = 0.00 < 0.05)
                        thrpt:  [+50.288% +52.328% +54.379%]
                        Performance has improved.
```